Moved generators to io.anserini.index.generator package.
