+++ src/main/java/io/anserini/index/DumpIndex.java
@@ -16,6 +16,7 @@
 
 package io.anserini.index;
 
+import io.anserini.index.generator.LuceneDocumentGenerator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.lucene.analysis.en.EnglishAnalyzer;
